+++ packages/storage/src/index.ts
@@ -1,3 +1,0 @@
-export * from './prisma/index.js';
-export * from './minio/index.js';
-export * from './mime/index.js';